requirements.txt
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn
plotly
surprise
implicit
azure-functions
azure-storage-blob
streamlit
